requirements.txt
aiodns==2.0.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==3.0.1
attrs==19.3.0
binance-futures-connector==4.1.0
brotlipy==0.7.0
certifi==2024.8.30
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.0
config==0.5.1
Django==5.1.4
django-scheduler==0.10.1
et_xmlfile==2.0.0
frozenlist==1.5.0
icalendar==6.1.0
idna==2.8
idna-ssl==1.1.0
multidict==6.1.0
numpy==2.1.3
openpyxl==3.1.5
pandas==2.2.3
propcache==0.2.1
pycares==4.5.0
pycparser==2.19
pycryptodome==3.21.0
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
schedule==1.2.2
six==1.16.0
sqlparse==0.5.2
tabulate==0.9.0
typing==3.7.4.1
typing-extensions==3.7.4.1
tzdata==2024.2
urllib3==2.2.3
websocket-client==1.8.0
yarl==1.18.3
